+++ src/test/java/com/github/alexisjehan/mavencheck/core/component/artifact/version/resolver/MavenArtifactAvailableVersionsResolverTest.java
@@ -160,6 +160,10 @@ final class MavenArtifactAvailableVersionsResolverTest {
 				.isThrownBy(
 						() -> mavenArtifactAvailableVersionsResolver.resolve(artifact, Collections.singletonList(null))
 				);
+		assertThatExceptionOfType(ArtifactAvailableVersionsResolveException.class)
+				.isThrownBy(
+						() -> mavenArtifactAvailableVersionsResolver.resolve(artifact, List.of())
+				);
 		assertThatExceptionOfType(AssertionError.class)
 				.isThrownBy(
 						() -> mavenArtifactAvailableVersionsResolver.resolve(artifact, repositories)