+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v1.2.0 (unreleased)
 
+### Improvements
+- Improve resolving artifact available versions by throwing an exception if no remote repository is available
+
 ## [v1.1.3](https://github.com/AlexisJehan/MavenCheck/releases/tag/v1.1.3) (2023-03-20)
 
 ### Notes